Fix bug where no event past 'updatestart' would be fired for a SourceBuffer.
If a GStreamer element (e.g. demuxer(s)) determines the stream is at the end, then no more buffers will be pulled from the source, and AppendPipeline will never see the endOfAppendBuffer.

We can fix this by hooking the eos signal on the sink and triggering the updateend event from there.